feat(ARCO-90): logger config
