+++ config/load.go
@@ -1,0 +1,49 @@
+package config
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/mitchellh/mapstructure"
+	"github.com/spf13/viper"
+)
+
+func Load(configFile string) (*ArcConfig, error) {
+	arcConfig := getDefaultArcConfig()
+
+	err := setDefaults(arcConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	viper.SetEnvPrefix("ARC")
+	viper.SetEnvKeyReplacer(strings.NewReplacer(".", "_"))
+	viper.AutomaticEnv()
+
+	viper.AddConfigPath(configFile)
+	err = viper.ReadInConfig()
+	if err != nil {
+		return nil, err
+	}
+
+	err = viper.Unmarshal(arcConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	return arcConfig, nil
+}
+
+func setDefaults(defaultConfig *ArcConfig) error {
+	defaultsMap := make(map[string]interface{})
+	if err := mapstructure.Decode(defaultConfig, &defaultsMap); err != nil {
+		err = fmt.Errorf("error occurred while setting defaults: %w", err)
+		return err
+	}
+
+	for key, value := range defaultsMap {
+		viper.SetDefault(key, value)
+	}
+
+	return nil
+}
